Extract useRecordVatEvents
